pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nba-betting-agent"
version = "0.1.0"
description = "Multi-agent NBA betting analysis system using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0.0",
    "langchain-core>=0.3.0",
    "typer[all]>=0.12.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
    "httpx[http2]>=0.28.0",
    "tenacity>=9.0.0",
    "circuitbreaker>=2.0.0",
    "nba_api>=1.4.0",
    "diskcache>=5.6.0",
    "scikit-learn>=1.4.0",
    "anthropic>=0.40.0",
    "langsmith>=0.2.0",
    "structlog>=25.5.0",
    "lightgbm>=4.0.0",
    "joblib>=1.3.0",
    "shap>=0.44.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.18.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "freezegun>=1.5.0",
    "pytest-freezegun>=0.4.2",
    "ollama>=0.4.0",
]
api = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
]

[project.scripts]
nba-ev = "nba_betting_agent.cli.main:main"
nba-dashboard = "nba_betting_agent.api.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nba_betting_agent*"]
